requirements.txt
Django==1.3
psycopg2==2.4.2
gunicorn==0.12.2
django-profiles==0.2
django-tagging==0.3.1
Pygments==1.4
Markdown==2.0.3
